# inflate: Read more bits when decoding