@klauspost 2020-02-19
Maybe @dsnet can spot some obvious flaw. After all he had the great idea to extend it to the size of EOB.

If not, this would be a reasonable thing to upstream.