@klauspost 2020-02-20
Tried a few variations on reading. Didn't manage to find anything faster.